requirements.txt
Django==3.2.4
django-smtp-ssl==1.0
pytz==2021.1
sqlparse==0.4.1
djangorestframework==*
